beginnings of unified search
